portfolio-deploy/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
llama-cpp-python==0.3.1
python-dotenv==1.0.0
pydantic==2.5.2
anthropic==0.8.1
#RAG
sentence-transformers==2.2.2
huggingface-hub==0.16.4   
transformers==4.33.2      
faiss-cpu==1.7.4
langchain==0.1.0
torch==2.0.1             